Added integration-tests for rating-service and some fix of RatingResponseDto

+++ rating-service/src/main/java/com/modsen/ratingservice/dto/response/RatingResponseDto.java
@@ -6,7 +6,7 @@ public record RatingResponseDto(
         Long id,
         String comment,
         String userType,
-        Integer refUserId,
+        Long refUserId,
         Integer rating,
         Long rideId) implements Serializable {
 }
